// src_frontend/api-client/models/LocationRoute.ts
/* tslint:disable */
/* eslint-disable */
/**
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface LocationRoute
 */
export interface LocationRoute {
  /**
   *
   * @type {string}
   * @memberof LocationRoute
   */
  id?: string;
  /**
   *
   * @type {Array<string>}
   * @memberof LocationRoute
   */
  readonly pickupLocationNames: Array<string>;
  /**
   *
   * @type {Date}
   * @memberof LocationRoute
   */
  readonly createdAt: Date;
  /**
   *
   * @type {Date}
   * @memberof LocationRoute
   */
  readonly updatedAt: Date;
  /**
   *
   * @type {string}
   * @memberof LocationRoute
   */
  name: string;
}

/**
 * Check if a given object implements the LocationRoute interface.
 */
export function instanceOfLocationRoute(value: object): value is LocationRoute {
  if (
    !("pickupLocationNames" in value) ||
    value["pickupLocationNames"] === undefined
  )
    return false;
  if (!("createdAt" in value) || value["createdAt"] === undefined) return false;
  if (!("updatedAt" in value) || value["updatedAt"] === undefined) return false;
  if (!("name" in value) || value["name"] === undefined) return false;
  return true;
}

export function LocationRouteFromJSON(json: any): LocationRoute {
  return LocationRouteFromJSONTyped(json, false);
}

export function LocationRouteFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): LocationRoute {
  if (json == null) {
    return json;
  }
  return {
    id: json["id"] == null ? undefined : json["id"],
    pickupLocationNames: json["pickup_location_names"],
    createdAt: new Date(json["created_at"]),
    updatedAt: new Date(json["updated_at"]),
    name: json["name"],
  };
}

export function LocationRouteToJSON(json: any): LocationRoute {
  return LocationRouteToJSONTyped(json, false);
}

export function LocationRouteToJSONTyped(
  value?: Omit<
    LocationRoute,
    "pickup_location_names" | "created_at" | "updated_at"
  > | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    id: value["id"],
    name: value["name"],
  };
}